Fix live timetable being empty at end-of-day
Before this fix, at the end of day, the initial comparison
that is searching for "Now" always falls into the return.

Because the "Now" item is in the past, and all the preview items for tomorrow
only have time (not date) and are thus in the future.

To fix this, detect the rollover between today and tomorrow.